+++ src/test/java/com/inq/wishhair/wesharewishhair/review/controller/utils/ReviewCreateRequestUtils.java
@@ -1,14 +1,10 @@
 package com.inq.wishhair.wesharewishhair.review.controller.utils;
 
 import com.inq.wishhair.wesharewishhair.global.fixture.ReviewFixture;
+import com.inq.wishhair.wesharewishhair.global.utils.MockMultipartFileUtils;
 import com.inq.wishhair.wesharewishhair.review.controller.dto.request.ReviewCreateRequest;
-import org.springframework.mock.web.MockMultipartFile;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.FileInputStream;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 
 
 public abstract class ReviewCreateRequestUtils {
@@ -18,22 +14,8 @@ public abstract class ReviewCreateRequestUtils {
         return new ReviewCreateRequest(
                 fixture.getContents(),
                 fixture.getScore(),
-                generateFiles(fixture),
+                MockMultipartFileUtils.generateFiles(fixture.getStoreUrls()),
                 hairStyleId
         );
-    }
-
-    private static List<MultipartFile> generateFiles(ReviewFixture fixture) throws IOException {
-        List<MultipartFile> files = new ArrayList<>();
-        for (String originalFilename : fixture.getStoreUrls()) {
-            files.add(createMultipartFile(originalFilename));
-        }
-        return files;
-    }
-
-    private static MultipartFile createMultipartFile(String originalFilename) throws IOException {
-        try (FileInputStream stream = new FileInputStream(FILE_PATH + originalFilename)) {
-            return new MockMultipartFile("files", originalFilename, "image/bmp", stream);
-        }
     }
 }
